requirements.txt
 os
 asyncio
 tornado
 speech_recognition 
 time 
 IPython.display as display